moving static time helpers to TimeUtils
