+++ app/src/main/java/co/touchlab/droidconandroid/shared/data/Event.java
@@ -156,12 +156,12 @@ public class Event implements ScheduleBlock
 
     public String getStartFormatted()
     {
-        return TimeUtils.DATE_FORMAT.get().format(new Date(startDateLong));
+        return TimeUtils.LOCAL_DATE_FORMAT.get().format(new Date(startDateLong));
     }
 
     public String getEndFormatted()
     {
-        return TimeUtils.DATE_FORMAT.get().format(new Date(endDateLong));
+        return TimeUtils.LOCAL_DATE_FORMAT.get().format(new Date(endDateLong));
     }
 
     public String getRsvpUuid()
