+++ app/src/main/java/co/touchlab/droidconandroid/shared/utils/TimeUtils.java
@@ -2,6 +2,7 @@ package co.touchlab.droidconandroid.shared.utils;
 import javax.annotation.Nonnull;
 
 import java.text.DateFormat;
+import java.text.ParseException;
 import java.text.SimpleDateFormat;
 import java.util.Calendar;
 import java.util.Date;
@@ -14,8 +15,13 @@ import java.util.TimeZone;
  */
 public class TimeUtils
 {
-    public static final TimeZone TIME_ZONE = TimeZone.getTimeZone("America/New_York");
-    public static ThreadLocal<DateFormat> DATE_FORMAT = new ThreadLocal<DateFormat>()
+    public static final SimpleDateFormat        SIMPLE_DATE_FORMAT = TimeUtils.makeDateFormat(
+            "MM/dd/yyyy");
+    public static final SimpleDateFormat        SIMPLE_TIME_FORMAT = TimeUtils.makeDateFormat(
+            "h:mma");
+    public static final TimeZone                TIME_ZONE          = TimeZone.getTimeZone(
+            "America/New_York");
+    public static       ThreadLocal<DateFormat> LOCAL_DATE_FORMAT  = new ThreadLocal<DateFormat>()
     {
         @Override
         protected DateFormat initialValue()
@@ -43,5 +49,15 @@ public class TimeUtils
         calendar.set(Calendar.SECOND, 0);                 // set second in minute
         calendar.set(Calendar.MILLISECOND, 0);
         return calendar.getTimeInMillis();
+    }
+
+    public static long parseTime(String date) throws ParseException
+    {
+        return TimeUtils.LOCAL_DATE_FORMAT.get().parse(date).getTime();
+    }
+
+    public static String dateToDayString(Date d)
+    {
+        return TimeUtils.SIMPLE_DATE_FORMAT.format(d);
     }
 }
